[Benchmark] Update Benchmarks with the Latest Quantization APIs (#44)
* [Update] Better interface for decoupled quantization

[Fix] remove auto transformation between shift and mul/div during simplification

[Add] pack API

* [Add] new tests for pack/unpack

* [Update] benchmarks with the latest quantization APIs

+++ heterocl/python/heterocl/api.py
@@ -441,7 +441,7 @@ def reduce_axis(min_, max_, name = "ra"):
   return _IterVar((min_, max_), name, 2)
 
 def reducer(init, freduce, dtype = "int32"):
-  def make_reduce(expr, axis, where = True, name = None, dtype = "int32"):
+  def make_reduce(expr, axis, where = True, name = None, dtype = dtype):
     if not isinstance(axis, (tuple, list)):
       axis = [axis]
     cb = CodeBuilder.current[-1]
